+++ LeetCode/345.cpp
@@ -14,7 +14,7 @@ private:
     }
 public:
     string reverseVowels(string s) {
-        int left = 0, right = s.length() - 1;
+        int left = 0, right = (int)s.length() - 1;
         while (left < right) {
             while (left < right && !isVowel(s[left])) left++;
             while (left < right && !isVowel(s[right])) right--;
